Move relocation utils to sentry.relocation
