+++ tests/sentry/relocation/api/endpoints/test_pause.py
@@ -10,9 +10,9 @@ from sentry.relocation.api.endpoints.pause import (
     ERR_NOT_PAUSABLE_STATUS,
 )
 from sentry.relocation.models.relocation import Relocation
+from sentry.relocation.utils import OrderedTask
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers.options import override_options
-from sentry.utils.relocation import OrderedTask
 
 TEST_DATE_ADDED = datetime(2023, 1, 23, 1, 23, 45, tzinfo=timezone.utc)
