+++ tests/sentry/relocation/api/endpoints/test_recover.py
@@ -12,9 +12,9 @@ from sentry.relocation.api.endpoints.recover import (
 )
 from sentry.relocation.models.relocation import Relocation
 from sentry.relocation.tasks import MAX_FAST_TASK_ATTEMPTS
+from sentry.relocation.utils import OrderedTask
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers.options import override_options
-from sentry.utils.relocation import OrderedTask
 
 TEST_DATE_ADDED = datetime(2023, 1, 23, 1, 23, 45, tzinfo=timezone.utc)
